Добавляет вынос функций и подключения к БД в отдельные файлы
