+++ config.ini
@@ -1,0 +1,6 @@
+;Настройки БД
+host=127.0.0.1
+user=root
+password=
+dbname=yeticave
+charset=utf8
